pyproject.toml
[project]
name = "agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "google-search-results>=2.4.2",
    "markdownify>=1.1.0",
    "mcp-server-fetch>=2025.4.7",
    "openai>=1.74.0",
    "openai-agents>=0.0.11",
    "selectolax>=0.3.28",
    "serpapi>=0.1.5",
    "smithery>=0.1.1",
]
